+++ wdist_dosage.c
@@ -1685,7 +1685,7 @@ int32_t oxford_distance_calc_unscanned(FILE* genfile, uint32_t* gen_buf_len_ptr,
   return retval;
 }
 
-int32_t wdist_dosage(int32_t calculation_type, int32_t dist_calc_type, char* genname, char* samplename, char* outname, char* outname_end, char* missing_code, int32_t distance_3d, int32_t distance_flat_missing, double exponent, int32_t maf_succ, uintptr_t regress_iters, uint32_t regress_d, uint32_t thread_ct, int32_t parallel_idx, uint32_t parallel_tot) {
+int32_t wdist_dosage(uint64_t calculation_type, int32_t dist_calc_type, char* genname, char* samplename, char* outname, char* outname_end, char* missing_code, int32_t distance_3d, int32_t distance_flat_missing, double exponent, int32_t maf_succ, uintptr_t regress_iters, uint32_t regress_d, uint32_t thread_ct, int32_t parallel_idx, uint32_t parallel_tot) {
   FILE* genfile = NULL;
   FILE* outfile = NULL;
   FILE* outfile2 = NULL;
